COMP: Consolodate duplicate LogTester code
ITK/Modules/Core/Common/test/itkLoggerThreadWrapperTest.cxx:100:7: warning: type ‘struct LogTester’ violates the C++ One Definition Rule [-Wodr]
 class LogTester
       ^
ITK/Modules/Core/Common/test/itkThreadLoggerTest.cxx:33:7: note: a different type is defined in another translation unit
 class LogTester
       ^
ITK/Modules/Core/Common/test/itkLoggerThreadWrapperTest.cxx:125:20: note: the first difference of corresponding definitions is field ‘m_Logger’
   itk::LoggerBase* m_Logger;
                    ^
