+++ Modules/Core/Common/test/itkLoggerTest.cxx
@@ -21,34 +21,7 @@
 #include "itkStdStreamLogOutput.h"
 #include "itkLogger.h"
 #include "itkTestingMacros.h"
-
-class LogTester
-{
-public:
-  LogTester(){ this->m_Logger = nullptr; }
-  itk::Logger* GetLogger() { return m_Logger; }
-  void SetLogger(itk::Logger* logger) { m_Logger = logger; }
-  void log() {
-    itkLogMacro( DEBUG, "DEBUG message by itkLogMacro\n" );
-    itkLogMacro( INFO, "INFO message by itkLogMacro\n" );
-    itkLogMacro( WARNING, "WARNING message by itkLogMacro\n" );
-    itkLogMacro( CRITICAL, "CRITICAL message by itkLogMacro\n" );
-    itkLogMacro( FATAL, "FATAL message by itkLogMacro\n" );
-    itkLogMacro( MUSTFLUSH, "MUSTFLUSH message by itkLogMacro\n" );
-  }
-  static void logStatic(LogTester* tester)
-  {
-    itkLogMacroStatic(tester, DEBUG, "DEBUG message by itkLogMacroStatic\n" );
-    itkLogMacroStatic(tester, INFO, "INFO message by itkLogMacroStatic\n" );
-    itkLogMacroStatic(tester, WARNING, "WARNING message by itkLogMacroStatic\n" );
-    itkLogMacroStatic(tester, CRITICAL, "CRITICAL message by itkLogMacroStatic\n" );
-    itkLogMacroStatic(tester, FATAL, "FATAL message by itkLogMacroStatic\n" );
-    itkLogMacroStatic(tester, MUSTFLUSH, "MUSTFLUSH message by itkLogMacroStatic\n" );
-  }
-
-private:
-  itk::Logger* m_Logger;
-};
+#include "itkLogTester.h"
 
 int itkLoggerTest( int argc, char *argv [] )
 {
@@ -88,12 +61,12 @@ int itkLoggerTest( int argc, char *argv [] )
 
     // Logging by the itkLogMacro from a class with itk::Logger
     std::cout << "  Logging by the itkLogMacro from a class with itk::Logger" << std::endl;
-    LogTester tester;
+    itk::Testing::LogTester tester;
     tester.SetLogger(logger);
     tester.log();
     // Logging by the itkLogMacroStatic from a class with itk::Logger
     std::cout << "  Logging by the itkLogMacroStatic from a class with itk::Logger" << std::endl;
-    LogTester::logStatic(&tester);
+    itk::Testing::LogTester::logStatic(&tester);
 
     // Writing by the logger
     std::cout << "  Writing by itk::Logger" << std::endl;
